Skip counter measurement when start listening while processing
